python-work/requirement.txt
# Core AI & ML
scikit-learn>=1.2.2
joblib>=1.2.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# NLP Libraries
spacy>=3.6.0
nltk>=3.8.1

# Resume Parsing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Data Visualization (Kaggle Notebook Dependencies)
matplotlib>=3.7.0
seaborn>=0.12.2
wordcloud>=1.9.3

# Web Framework (for API Integration later)
Flask>=2.3.0
fastapi>=0.110.0
uvicorn>=0.23.0

# Additional Utilities
requests>=2.31.0
